Moved generate_embeddings_index to common.py from model file
