+++ library/common.py
@@ -27,6 +27,17 @@ def get_sequence_of_tokens(corpus):
     return input_sequences, total_words
 
 
+# generate embeddings index
+def generate_embeddings_index(GLOVE_DIR):
+    embeddings_index = {}
+    with open(os.path.join(GLOVE_DIR, 'glove.6B.100d.txt')) as f:
+        for line in f:
+            word, coefs = line.split(maxsplit=1)
+            coefs = np.fromstring(coefs, 'f', sep=' ')
+            embeddings_index[word] = coefs
+    return embeddings_index
+
+
 # return word index
 def get_word_index(corpus):
     tokenizer = Tokenizer()
